update to use context for view management
