+++ sales-coach-avatar/src/app/_views/index.ts
@@ -1,2 +1,3 @@
 export * from "./demo";
 export * from "./description";
+export * from './initial';
