+++ sales-coach-avatar/src/app/_views/initial/index.ts
@@ -1,0 +1,1 @@
+export * from './initial';
